Migrated overwrite-time and utils to separate files
