+++ index.js
@@ -31,224 +31,13 @@
  */
 
 const puppeteer = require('puppeteer');
-const fs = require('fs');
 const path = require('path');
 const sprintf = require('sprintf-js').sprintf;
 const defaultDuration = 5;
 const defaultFPS = 60;
 const { overwriteRandom } = require('./lib/overwrite-random');
-
-const overwriteTime = function (page, animationFrameDuration) {
-  return page.evaluateOnNewDocument(function (animationFrameDuration) {
-    (function (exports) {
-      var _virtualTime = (new Date()).getTime();
-      var _startTime = _virtualTime;
-      var _oldDate = Date;
-      // a block is a segment of blocking code, wrapped in a function
-      // to be run at a certain virtual time. They're created by
-      // window.requestAnimationFrame, window.setTimeout, and window.setInterval
-      var _pendingBlocks = [];
-      var _intervals = {};
-      var _idCount = 1;
-      var _sortPendingBlocks = function () {
-        _pendingBlocks = _pendingBlocks.sort(function (a, b) {
-          if (a.time !== b.time) {
-            return a.time - b.time;
-          }
-          return a.id - b.id;
-        });
-      };
-      var _processNextBlock = function () {
-        if (!_pendingBlocks.length) {
-          return null;
-        }
-        _sortPendingBlocks();
-        var block = _pendingBlocks.shift();
-        _virtualTime = block.time;
-        block.fn.apply(exports, block.args);
-      };
-      var _processUntilTime = function (ms) {
-        _sortPendingBlocks();
-
-        while (_pendingBlocks.length && _pendingBlocks[0].time <= _startTime + ms) {
-          _processNextBlock();
-          _sortPendingBlocks();
-        }
-        // TODO: maybe wait a little while for possible promises to resolve?
-        _virtualTime = _startTime + ms;
-      };
-      var _setTimeout = function (fn, timeout, ...args) {
-        var id = _idCount;
-        var blockFn;
-        if (fn instanceof Function) {
-          blockFn = fn;
-        } else {
-          // according to https://developer.mozilla.org/en-US/docs/Web/API/WindowOrWorkerGlobalScope/setTimeout,
-          // setTimeout should support evaluating strings as code, though it's not recommended
-          blockFn = function () {
-            eval(fn);
-          };
-        }
-        if (!timeout || isNaN(timeout)) {
-          // If timeout is 0, there may be an infinite loop
-          // Changing it to 1 shouldn't disrupt code, because
-          // setTimeout doesn't usually execute code immediately
-          timeout = 1;
-        }
-        _pendingBlocks.push({
-          time: timeout + _virtualTime,
-          id: id,
-          fn: blockFn,
-          args: args
-        });
-        _idCount++;
-        return id;
-      };
-
-      var _clearTimeout = function (id) {
-        // according to https://developer.mozilla.org/en-US-docs/Web/API/WindowOrWorkerGlobalScope/setInterval,
-        // setInterval and setTimeout share the same pool of IDs, and clearInterval and clearTimeout
-        // can technically be used interchangeably
-        var i = 0;
-        if (_intervals[id]) {
-          _intervals[id].clear();
-        }
-        while (i < _pendingBlocks.length) {
-          if (_pendingBlocks[i].id === id) {
-            _pendingBlocks.splice(i, 1);
-          } else {
-            i++;
-          }
-        }
-      };
-
-      var _frameTime;
-      var _requestAnimationFrame = function (fn) {
-        return _setTimeout(function () {
-          // According to https://developer.mozilla.org/en-US/docs/Web/API/window/requestAnimationFrame,
-          // the passed argument to the callback should be the starting time of the
-          // chunk of requestAnimationFrame callbacks that are called for that particular frame.
-          // Since the processing time of callbacks do not advance virtual time, in most cases
-          // there may not be significant differences between _frameTime and _virtualTime.
-          fn(_frameTime);
-        }, animationFrameDuration);
-      };
-
-      var _updateFrameTime = function () {
-        // Using this implementation may potentially cause issues in the future
-        // for high-fps capture, where _frameTime is not advanced per frame
-        _requestAnimationFrame(_updateFrameTime);
-        _frameTime = _virtualTime;
-      };
-      _updateFrameTime();
-
-      // overwriting built-in functions...
-      exports.Date = class Date extends _oldDate {
-        constructor() {
-          if (!arguments.length) {
-            super(_virtualTime);
-          } else {
-            super(...arguments);
-          }
-        }
-      };
-      exports.Date.now = exports.performance.now = function () {
-        return _virtualTime;
-      };
-      exports.setTimeout = _setTimeout;
-      exports.requestAnimationFrame = _requestAnimationFrame;
-      exports.setInterval = function (fn, interval, ...args) {
-        var lastCallId;
-        var id = _idCount;
-        var running = true;
-        var intervalFn = function () {
-          if (fn instanceof Function) {
-            fn.apply(exports, args);
-          } else {
-            // according to https://developer.mozilla.org/en-US/docs/Web/API/WindowOrWorkerGlobalScope/setInterval,
-            // setInterval should support evaluating strings as code, though it's not recommended
-            eval(fn);
-          }
-          if (running) {
-            lastCallId = _setTimeout(intervalFn, interval);
-          }
-        };
-        _intervals[id] = {
-          clear: function () {
-            _clearTimeout(lastCallId);
-            running = false;
-          }
-        };
-        _idCount++;
-        lastCallId = _setTimeout(intervalFn, interval);
-        // according to https://developer.mozilla.org/en-US-docs/Web/API/WindowOrWorkerGlobalScope/setInterval,
-        // setInterval and setTimeout share the same pool of IDs, and clearInterval and clearTimeout
-        // can technically be used interchangeably
-        return id;
-      };
-      exports.cancelAnimationFrame = _clearTimeout;
-      exports.clearTimeout = _clearTimeout;
-      exports.clearInterval = _clearTimeout;
-      // exported custom functions
-      exports._processNextBlock = _processNextBlock;
-      exports._processUntilTime  = _processUntilTime;
-    })(this);
-  }, animationFrameDuration);
-};
-
-const promiseLoop = function (condition, body) {
-  var loop = function () {
-    if (condition()) {
-      return body().then(loop);
-    }
-  };
-  return Promise.resolve().then(loop);
-};
-
-const getBrowserFrames = function (frame) {
-  return [frame].concat(...frame.childFrames().map(getBrowserFrames));
-};
-
-const goToTime = function (browserFrames, time) {
-  // Goes to a certain time. Can't go backwards
-  return Promise.all(browserFrames.map(function (frame) {
-    return frame.evaluate(function (ms) {
-      window._processUntilTime(ms);
-    }, time);
-  }));
-};
-
-const getSelectorDimensions = function (page, selector) {
-  return page.evaluate(function (selector) {
-    var el = document.querySelector(selector);
-    var dim = el.getBoundingClientRect();
-    if (el) {
-      return {
-        left: dim.left,
-        top: dim.top,
-        right: dim.right,
-        bottom: dim.bottom,
-        scrollX: window.scrollX,
-        scrollY: window.scrollY,
-        x: dim.x,
-        y: dim.y,
-        width: dim.width,
-        height: dim.height
-      };
-    }
-  }, selector);
-};
-
-const makeFileDirectoryIfNeeded = function (filepath) {
-  var dir = path.parse(filepath).dir, ind, currDir;
-  var directories = dir.split(path.sep);
-  for (ind = 1; ind <= directories.length; ind++) {
-    currDir = directories.slice(0, ind).join(path.sep);
-    if (currDir && !fs.existsSync(currDir)) {
-      fs.mkdirSync(currDir);
-    }
-  }
-};
+const { overwriteTime, goToTime } = require('./lib/overwrite-time');
+const { promiseLoop, getBrowserFrames, getSelectorDimensions, makeFileDirectoryIfNeeded } = require('./lib/utils');
 
 module.exports = function (config) {
   config = Object.assign({}, config || {});
